+++ etc/input.conf
@@ -192,19 +192,19 @@
 #JOY_BTN2 add volume 1
 #JOY_BTN3 add volume -1
 
-# For dvdnav://
+# For dvdnav:// and bdnav://
 
-# dvdnav controls during playback
-#ENTER  {dvdnav} dvdnav menu        # DVDNav MENU
-# BS    {dvdnav} dvdnav prev        # DVDNav PREVIOUS menu (in the order chapter->title->root)
-# dvdnav controls when showing menu (additionally to the controls above)
-#UP    {dvdnav-menu} dvdnav up          # DVDNav UP
-#DOWN  {dvdnav-menu} dvdnav down        # DVDNav DOWN
-#LEFT  {dvdnav-menu} dvdnav left        # DVDNav LEFT
-#RIGHT {dvdnav-menu} dvdnav right       # DVDNav RIGHT
-#ENTER {dvdnav-menu} dvdnav select      # DVDNav SELECT (ok)
-#MOUSE_BTN0 {dvdnav-menu} dvdnav mouse
-#MOUSE_MOVE {dvdnav-menu} dvdnav mouse_move
+# navigation controls during playback
+#ENTER  {discnav} discnav menu        # DISCNAV MENU
+# BS    {discnav} discnav prev        # DISCNAV PREVIOUS menu (in the order chapter->title->root)
+# navigation controls when showing menu (additionally to the controls above)
+#UP    {discnav-menu} discnav up          # DISCNAV UP
+#DOWN  {discnav-menu} discnav down        # DISCNAV DOWN
+#LEFT  {discnav-menu} discnav left        # DISCNAV LEFT
+#RIGHT {discnav-menu} discnav right       # DISCNAV RIGHT
+#ENTER {discnav-menu} discnav select      # DISCNAV SELECT (ok)
+#MOUSE_BTN0 {discnav-menu} discnav mouse
+#MOUSE_MOVE {discnav-menu} discnav mouse_move
 
 #
 # Not assigned by default
