+++ old-makefile
@@ -222,7 +222,7 @@ SOURCES = audio/audio.c \
           player/client.c \
           player/configfiles.c \
           player/command.c \
-          player/dvdnav.c \
+          player/discnav.c \
           player/loadfile.c \
           player/main.c \
           player/misc.c \
