+++ wscript_build.py
@@ -209,7 +209,7 @@ def build(ctx):
         ( "player/client.c" ),
         ( "player/command.c" ),
         ( "player/configfiles.c" ),
-        ( "player/dvdnav.c" ),
+        ( "player/discnav.c" ),
         ( "player/loadfile.c" ),
         ( "player/main.c" ),
         ( "player/misc.c" ),
